Fix broken logic for checking bounds
Below the two uint32's are added together, and _then_ assigned to a
quint64, and then it is checked if it is too large for a uint32. So
instead make sure we get in a proper large number, avoids casting.

+++ src/libs/ssh/sshchannel.cpp
@@ -103,7 +103,7 @@ quint32 AbstractSshChannel::maxPacketSize()
     return 16 * 1024 * 1024;
 }
 
-void AbstractSshChannel::handleWindowAdjust(quint32 bytesToAdd)
+void AbstractSshChannel::handleWindowAdjust(quint64 bytesToAdd)
 {
     checkChannelActive();
